DT: bcm2712: override supports-cqe to a cell
We want to be able to control the interop surface exposed by Command
Queueing across bcm2712 products to a more restrictive default, with
selectable disable and permissive behaviour.

Changing the bool to a cell lets it relay a tristate value.

Also add the override parameter to CM5 as CM5-lite may interface with
arbitrary eMMC or SD cards.

(Reimplemented on 6.12 - bcm2712 dts now has a downstream/upstream split.)

Signed-off-by: Jonathan Bell <jonathan@raspberrypi.com>